refactoring(plugins): move code related to manifest of managed plugins to a dedicated module
Follows-up https://github.com/qgis-deployment/qgis-deployment-toolbelt-cli/pull/875#issuecomment-5131701760
